# Giorno 13/api.spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gestione Studenti API",
    "description": "API CRUD per la gestione dell'anagrafica studenti",
    "contact": {
      "name": "Lorenzo Sansone",
      "email": "lorenzo.sansone@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Studenti",
      "description": "Gestione completa dell'anagrafica studenti"
    }
  ],
  "paths": {
    "/api/studenti": {
      "get": {
        "tags": [
          "Studenti"
        ],
        "summary": "Recupera la lista paginata di tutti gli studenti",
        "operationId": "getAll",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista recuperata con successo",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageStudente"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Studenti"
        ],
        "summary": "Crea un nuovo record studente",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Studente"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Studente creato con successo",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Studente"
                }
              }
            }
          },
          "400": {
            "description": "Dati forniti non validi",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Studente"
                }
              }
            }
          }
        }
      }
    },
    "/api/studenti/{id}": {
      "get": {
        "tags": [
          "Studenti"
        ],
        "summary": "Recupera un singolo studente tramite il suo ID",
        "operationId": "getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Studente trovato",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Studente"
                }
              }
            }
          },
          "404": {
            "description": "Studente non trovato",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Studente"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Studenti"
        ],
        "summary": "Elimina uno studente dal database tramite ID",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Studente eliminato con successo"
          },
          "404": {
            "description": "Studente non trovato"
          }
        }
      },
      "patch": {
        "tags": [
          "Studenti"
        ],
        "summary": "Aggiorna parzialmente i dati di uno studente esistente",
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "object"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Studente aggiornato",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Studente"
                }
              }
            }
          },
          "404": {
            "description": "Studente non trovato",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Studente"
                }
              }
            }
          },
          "400": {
            "description": "Modifiche non valide",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Studente"
                }
              }
            }
          }
        }
      }
    },
    "/api/studenti/{id}/corso": {
      "get": {
        "tags": [
          "Studenti"
        ],
        "summary": "Recupera solo il nome del corso di laurea di uno studente specifico",
        "operationId": "getCorso",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Corso recuperato",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Studente non trovato",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/studenti/corso/{corso}": {
      "get": {
        "tags": [
          "Studenti"
        ],
        "summary": "Cerca studenti iscritti a un determinato corso di laurea",
        "operationId": "getByCorsoLaurea",
        "parameters": [
          {
            "name": "corso",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ricerca completata",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Studente"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/studenti/cognome/{cognome}": {
      "get": {
        "tags": [
          "Studenti"
        ],
        "summary": "Cerca studenti filtrando per cognome",
        "operationId": "getByCognomePath",
        "parameters": [
          {
            "name": "cognome",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ricerca completata (può restituire una lista vuota)",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Studente"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Studente": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "ID univoco dello studente",
            "example": 1,
            "readOnly": true
          },
          "nome": {
            "type": "string",
            "description": "Nome dello studente",
            "example": "Carlo"
          },
          "cognome": {
            "type": "string",
            "description": "Cognome dello studente",
            "example": "Prestigiacomo"
          },
          "email": {
            "type": "string",
            "description": "Indirizzo email dello studente",
            "example": "carlo.prestigiacomo@gmail.com"
          },
          "dataNascita": {
            "type": "string",
            "format": "date",
            "description": "Data di nascita dello studente",
            "example": "2005-11-17"
          },
          "corsoLaurea": {
            "type": "string",
            "description": "Corso di laurea a cui lo studente è iscritto",
            "example": "Informatica"
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageStudente": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Studente"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          },
          "paged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      }
    }
  }
}